+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# v3.5.0
+## 08/31/2021
+
+1. [](#new)
+   * Require **Grav 1.7.19**, **Form 5.1.0** and **Email 3.1.0**
+   * Added support for adding login pages by `$grav['login']->addPage()`
+   * Added support for getting all login routes by `$grav['login']->getRoute()`
+   * Added support for form layouts
+   * Deprecated methods `LoginPlugin::defaultRedirectAfterLogin()` and `LoginPlugin::defaultRedirectAfterLogout()`
+1. [](#improved)
+   * Speed up `visibility_requires_access` checks by restricting full check to only visible pages
+1. [](#bugfix)
+   * Fixed login pages with redirect breaking the site
+
 # v3.4.4
 ## 06/16/2021